Get rid of hooks for filter registration
