+++ domains/eventEditor/src/services/filterState/datetimes/index.ts
@@ -1,7 +1,5 @@
 export { default as useDatesListFilterState } from './useDatesListFilterState';
 
-export { default as useDatesFilterBarService } from './useDatesFilterBarService';
-
 export { default as useDatesListFilterStateManager } from './useDatesListFilterStateManager';
 
 export * from './types';
