+++ domains/eventEditor/src/services/filterState/registerFilters.ts
@@ -1,0 +1,2 @@
+import './datetimes/registerFilters';
+import './tickets/registerFilters';
